fix(examples): make example-code typecheck — promote createDefaultSessionState, drop ghost tsconfig extends
the eliza-code restore left develop's example-code red three ways: an
unformatted entry file (previous commit), an import of
createDefaultSessionState from lib/session that only ever existed as a
private copy in cli.ts, and a tsconfig extends of tsconfig.dist-paths.json
which the workspace consolidation deleted. promote the session-state
factory into lib/session as the single shared export, import it from
cli.ts, and drop the ghost extends (the package's own compilerOptions with
bundler resolution cover it).

verified: bun run typecheck and lint:check both green in the package.
