+++ packages/examples/code/tsconfig.json
@@ -21,6 +21,5 @@
     "types": ["node", "bun"]
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"],
-  "extends": "../../../tsconfig.dist-paths.json"
+  "exclude": ["node_modules", "dist"]
 }
